Add "meaninglessFileNames" option
The "meaninglessFileNames" option provides the user with more precise
control over the component display name.

This option is to be used in conjunction with "displayName" and
"fileName". If either is set to false, this option will have no effect.

Prior to this commit, when both "displayName" and "fileName" were set,
the behaviour was to prepend the file name to the component display name
if it was named anything other than "index", and to prepend the
directory name otherwise.

The "meaninglessFileNames" option enables developers to control when
exactly the directory name should be used instead of the file name. If
the file name is considered to be "meaningless", that is, it doesn't
provide the developer with any useful information, then the directory
name will be used instead.

By default, the only "meaningless file name" is "index", which means
that the default behaviour is unmodified.